JWWEB-737 Divined different status items to different files
