+++ .idea/inspectionProfiles/Project_Default.xml
@@ -7,6 +7,9 @@
     <inspection_tool class="ElmUnusedImport" enabled="false" level="WARNING" enabled_by_default="false" />
     <inspection_tool class="ElmUnusedSymbol" enabled="false" level="WARNING" enabled_by_default="false" />
     <inspection_tool class="Eslint" enabled="true" level="ERROR" enabled_by_default="true" />
+    <inspection_tool class="JSConstructorReturnsPrimitive" enabled="true" level="WEAK WARNING" enabled_by_default="true">
+      <scope name="Project Files" level="Notice" enabled="true" />
+    </inspection_tool>
     <inspection_tool class="Stylelint" enabled="true" level="ERROR" enabled_by_default="true" />
     <inspection_tool class="UnterminatedStatementJS" enabled="false" level="WARNING" enabled_by_default="false">
       <option name="ignoreSemicolonAtEndOfBlock" value="true" />
