+++ .idea/inspectionProfiles/profiles_settings.xml
@@ -1,0 +1,26 @@
+<component name="InspectionProjectProfileManager">
+  <settings>
+    <info color="5bbe5e">
+      <option name="EFFECT_COLOR" value="5bbe5e" />
+      <option name="ERROR_STRIPE_COLOR" value="5bbe5e" />
+      <option name="EFFECT_TYPE" value="2" />
+      <option name="myName" value="Notice" />
+      <option name="myVal" value="50" />
+      <option name="myExternalName" value="Notice" />
+      <option name="myDefaultAttributes">
+        <option name="ERROR_STRIPE_COLOR" value="5bbe5e" />
+      </option>
+    </info>
+    <list size="9">
+      <item index="0" class="java.lang.String" itemvalue="INFORMATION" />
+      <item index="1" class="java.lang.String" itemvalue="Notice" />
+      <item index="2" class="java.lang.String" itemvalue="UNUSED ENTRY" />
+      <item index="3" class="java.lang.String" itemvalue="TYPO" />
+      <item index="4" class="java.lang.String" itemvalue="SERVER PROBLEM" />
+      <item index="5" class="java.lang.String" itemvalue="WEAK WARNING" />
+      <item index="6" class="java.lang.String" itemvalue="INFO" />
+      <item index="7" class="java.lang.String" itemvalue="WARNING" />
+      <item index="8" class="java.lang.String" itemvalue="ERROR" />
+    </list>
+  </settings>
+</component>
