+++ src/components/TransactionDetails/templates/index.js
@@ -1,0 +1,4 @@
+// @flow strict
+
+export { TransactionNormalTemplate } from './TransactionNormal'
+export { TransactionCancelTemplate } from './TransactionCancel'
