+++ src/components/TransactionDetails/utils/index.js
@@ -1,0 +1,3 @@
+// @flow strict
+
+export { handleEditNote } from './handleEditNote'
